new method for calculating cursor matches, fancier graphics

+++ src/cursor.cpp
@@ -53,7 +53,7 @@ void Cursor::set_mouse_position(int x, int y)
 {
 	m_mouse_x = x;
 	m_mouse_y = y;
-	m_sprite.setPosition(m_mouse_x-m_cursorsize/2, m_mouse_y-m_cursorsize/2);
+	m_sprite.setPosition((int)(m_mouse_x-m_cursorsize/2.0), (int)(m_mouse_y-m_cursorsize/2.0));
 }
 int Cursor::get_mouse_position_x(void)
 {
@@ -66,7 +66,7 @@ int Cursor::get_mouse_position_y(void)
 int Cursor::calculate_sizes(int w, int h)
 {
 	m_cursorsize = h*(SIZE_CURSOR/100.0);
-	m_sprite.setScale(m_cursorsize/16.0, m_cursorsize/16.0);
+	m_sprite.setScale((int)(m_cursorsize/16.0), (int)(m_cursorsize/16.0));
 	return 0;
 }
 sf::Sprite Cursor::get_drawable(int w, int h)
